CMenuWidget: add missing hide() for details line
Details_line was not removed in all constellations when menu was closed.
Sample: open main menue and select an entry with connect-line, eg. plugins,
press key to come into submenu, then go back to main menue and close
menu with home key, result: it's possible that line is still on screen.