hdd_menu.cpp: remove "-T largefile" parameter
Some users experience problems with too low amount of inodes (No space
left on device but only a few percent of the device are actually used).
To fix this, simply use default settings. This will give us a much
higher amount of inodes on the hard drive but will waste a bit more
space on the device for inode reservation.