osd_setup/menue.cpp: remove slot OnAfterChangeOption from OsdSetup class
Effect was right only for the osd_menu widget but not for all the
other opened menu widget instances. This was watched if user the osd setup menu
has lefted with the 'Setup' ('Menu') button. The main menu was not touched.
Thx GetAway for hint.
The use of slot in context with the OsdHelpers class should solve this
issue for all new generated menu widget instances.