movieplayer: add send message CRCInput::RC_info on changed position
Sends message to show info bar on change position, maybe there is
a better solution (e.g. with callback), but this should help to visualize if position
was changed, e.g.: user has pressed number keys for minute steps.
In my opinion, this feature is missing for better user interaction.