display: add class CDisplay
This should fix broken build with --enable-lcd (BOXMODEL_OSMIO4KPLUS)

FIXME: CLCD::setVolume() needs fix because it contains no code.

The concept to overwrite CVFD with CLCD should be thought over, see:

display.h:

#if BOXMODEL_OSMIO4KPLUS
#include <driver/lcdd.h>
#define CVFD CLCD
#else
#include <driver/simple_display.h>
#endif

A plan with base- and subclasses and overwritable members should be
more sustainable. It will be coming to a mess, and if not yet, it will
be at some point, especially since nothing was documented.